+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,4 @@
+# 課題
+
+# やること
+
